Fix stale item ignoring
Action.Ignore has a specific meaning - the new item data is copied to
the existing one but it doesn't move on the list (might be filtered out,
though). When a stale item comes in, we want to ignore it complete, so
add a Action.Ignore flag for that. Also, make the newer comparer not
mandatory.